requirements.txt
alembic==1.7.1
amqp==5.0.6
anyio==3.1.0
appdirs==1.4.4
asgi-lifespan==1.0.1
asgiref==3.3.4
astroid==2.4.2
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.1.0
autopep8==1.5.4
base58==2.0.1
bcrypt==3.2.0
billiard==3.6.4.0
bitarray==1.2.2
bson==0.5.10
cached-property==1.5.2
cachetools==4.2.2
case-converter==1.0.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
contextvars==2.4
cssselect==1.1.0
cssutils==2.3.0
cytoolz==0.10.1
dataclasses==0.6
distlib==0.3.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.2
emails==0.6
eth-abi==2.1.1
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.4
fastapi==0.68.2
filelock==3.0.12
https://github.com/blaize-tech/celo-sdk-py/archive/32bb8f2.zip
greenlet==1.0.0
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.2.1
httpcore==0.13.3
httptools==0.2.0
httpx==0.18.1
idna==2.10
immutables==0.15
importlib-metadata==4.6.1
importlib-resources==5.2.0
iniconfig==1.1.1
ipfshttpclient==0.6.0.post1
isort==5.4.2
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.1.0
lazy-object-proxy==1.4.3
limits==1.5.1
lru-dict==1.1.6
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
mirakuru==2.3.0
motor==2.5.1
multiaddr==0.0.9
netaddr==0.8.0
nums-from-string==0.1.2
packaging==20.9
parsimonious==0.8.1
passlib==1.7.4
pluggy==0.13.1
port-for==0.4
premailer==3.9.0
prompt-toolkit==3.0.18
protobuf==3.13.0
psutil==5.8.0
py==1.10.0
py-solc==3.2.0
py-solc-x==0.10.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pydantic==1.8.2
pylint==2.6.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.3
pytest-asyncio==0.16.0
pytest-env==0.6.2
pytest-mock==3.6.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
raven==6.10.0
requests==2.24.0
rfc3986==1.5.0
rlp==1.2.0
rsa==4.7.2
semantic-version==2.8.5
six==1.15.0
sniffio==1.2.0
solc==0.0.0a0
starlette==0.14.2
text-unidecode==1.3
toml==0.10.1
toolz==0.10.0
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.14.0
uvloop==0.15.2
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.12.0
websockets==8.1
wrapt==1.12.1
zipp==3.5.0
